Reorganize normalized
