+++ sdk/typescript/src/types/objects.ts
@@ -69,87 +69,6 @@ export type SuiData = Infer<typeof SuiData>;
 
 export const MIST_PER_SUI = BigInt(1000000000);
 
-export type SuiMoveFunctionArgTypesResponse = Infer<
-  typeof SuiMoveFunctionArgType
->[];
-
-export const SuiMoveFunctionArgType = union([
-  string(),
-  object({ Object: string() }),
-]);
-
-export const SuiMoveFunctionArgTypes = array(SuiMoveFunctionArgType);
-export type SuiMoveFunctionArgTypes = Infer<typeof SuiMoveFunctionArgTypes>;
-
-export type SuiMoveNormalizedModules = Record<string, SuiMoveNormalizedModule>;
-
-export type SuiMoveNormalizedModule = {
-  file_format_version: number;
-  address: string;
-  name: string;
-  friends: SuiMoveModuleId[];
-  structs: Record<string, SuiMoveNormalizedStruct>;
-  exposed_functions: Record<string, SuiMoveNormalizedFunction>;
-};
-
-export type SuiMoveModuleId = {
-  address: string;
-  name: string;
-};
-
-export type SuiMoveNormalizedStruct = {
-  abilities: SuiMoveAbilitySet;
-  type_parameters: SuiMoveStructTypeParameter[];
-  fields: SuiMoveNormalizedField[];
-};
-
-export type SuiMoveStructTypeParameter = {
-  constraints: SuiMoveAbilitySet;
-  is_phantom: boolean;
-};
-
-export type SuiMoveNormalizedField = {
-  name: string;
-  type_: SuiMoveNormalizedType;
-};
-
-export type SuiMoveNormalizedFunction = {
-  visibility: SuiMoveVisibility;
-  is_entry: boolean;
-  type_parameters: SuiMoveAbilitySet[];
-  parameters: SuiMoveNormalizedType[];
-  return_: SuiMoveNormalizedType[];
-};
-
-export type SuiMoveVisibility = 'Private' | 'Public' | 'Friend';
-
-export type SuiMoveTypeParameterIndex = number;
-
-export type SuiMoveAbilitySet = {
-  abilities: string[];
-};
-
-export type SuiMoveNormalizedType =
-  | string
-  | SuiMoveNormalizedTypeParameterType
-  | { Reference: SuiMoveNormalizedType }
-  | { MutableReference: SuiMoveNormalizedType }
-  | { Vector: SuiMoveNormalizedType }
-  | SuiMoveNormalizedStructType;
-
-export type SuiMoveNormalizedTypeParameterType = {
-  TypeParameter: SuiMoveTypeParameterIndex;
-};
-
-export type SuiMoveNormalizedStructType = {
-  Struct: {
-    address: string;
-    module: string;
-    name: string;
-    type_arguments: SuiMoveNormalizedType[];
-  };
-};
-
 export const SuiObject = object({
   /** The meat of the object */
   data: SuiData,
@@ -323,41 +242,4 @@ export function getMovePackageContent(
     return undefined;
   }
   return (suiObject.data as SuiMovePackage).disassembled;
-}
-
-export function extractMutableReference(
-  normalizedType: SuiMoveNormalizedType
-): SuiMoveNormalizedType | undefined {
-  return typeof normalizedType === 'object' &&
-    'MutableReference' in normalizedType
-    ? normalizedType.MutableReference
-    : undefined;
-}
-
-export function extractReference(
-  normalizedType: SuiMoveNormalizedType
-): SuiMoveNormalizedType | undefined {
-  return typeof normalizedType === 'object' && 'Reference' in normalizedType
-    ? normalizedType.Reference
-    : undefined;
-}
-
-export function extractStructTag(
-  normalizedType: SuiMoveNormalizedType
-): SuiMoveNormalizedStructType | undefined {
-  if (typeof normalizedType === 'object' && 'Struct' in normalizedType) {
-    return normalizedType;
-  }
-
-  const ref = extractReference(normalizedType);
-  const mutRef = extractMutableReference(normalizedType);
-
-  if (typeof ref === 'object' && 'Struct' in ref) {
-    return ref;
-  }
-
-  if (typeof mutRef === 'object' && 'Struct' in mutRef) {
-    return mutRef;
-  }
-  return undefined;
 }
